add a alart to take user permisson for user current location
