+++ public/locals/bn/translation.json
@@ -1,5 +1,6 @@
 {
   "home": "হোম",
+  "memberships":"সদস্যপদ",
   "services": "সেবা",
   "about": "আমাদের সম্পর্কে",
   "contact": "যোগাযোগ",
