+++ public/locals/en/translation.json
@@ -1,5 +1,6 @@
 {
   "home": "Home",
+  "memberships":"Memberships",
   "services": "Services",
   "about": "About",
   "contact": "Contact",
